refactor : add Lotto
