+++ src/test/java/lotto/LottoTest.java
@@ -1,5 +1,7 @@
-package calculator;
+package lotto;
 
+import calculator.PositiveNumber;
+import lotto.domain.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -14,19 +16,11 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatExceptionOf
  */
 public class LottoTest {
 
-    class Lotto {
-        private final List<PositiveNumber> numbers;
-
-        public Lotto(List<PositiveNumber> numbers) {
-            this.numbers = numbers;
-        }
-    }
-
     @Test
     @DisplayName("로또 생성")
     void create() {
         List<PositiveNumber> numbers = Arrays.asList(new PositiveNumber("8"), new PositiveNumber("23"), new PositiveNumber("41"), new PositiveNumber("42"), new PositiveNumber("43"));
         Lotto l = new Lotto(numbers);
-        assertThat(l.numbers.size()).isEqualTo(5);
+        assertThat(l.getNumbers().size()).isEqualTo(5);
     }
 }
